arrumando todos os bugs falta relacionar pagamento e jogador

+++ src/main/java/dw/Mensalistas/repository/JogadorRepository.java
@@ -6,9 +6,9 @@ import dw.Mensalistas.model.Jogador;
 
 import java.util.List;
 
-public interface JogadorRepository extends JpaRepository<Jogador, Long>{
+public interface JogadorRepository extends JpaRepository<Jogador, Integer>{
     //List<Jogador> findByPublicado(boolean publicado);
-    //List<Jogador> findByCodJogador(int cod_jogador);
+    //Jogador> findByCodJogador(int cod_jogador);
     List<Jogador> findByNomeContaining(String nome);
 
 }
